Highest black pixel by sections

Break the original matrix in a number of section ( n_sections), then find the higher black pixel in each image section.

altitudinal_profile(imagematrix, n_sections, height_size)

Arguments

imagematrix
The matrix to be analysed.

n_sections
Break the image in this number of columns.

height_size
Real size of image height (in mm, cm, m, etc..).

Value

Mean

Height mean of the highest black pixel in sections.

SD

Standard deviations of the highest black pixel in sections.

Size

Height of the highest black pixel in sections.

References

Zehm et al 2003 Multiparameter analysis of vertical vegetation structure based on digital image processing. Flora-Morphology, Distribution, Functional Ecology of Plants, 198: 142-160.

See also

Examples

# First, get a matrix from your image. Here an example of a bush image is used.
bush<-system.file("extdata/bush.JPG",package ="bwimage")
bush_imagematrix<-threshold_color(bush,  "jpeg",  "proportional", compress_rate = 0.1)

# Profile  of  highest black pixels on sections of the bush image matrix
altitudinal_profile(bush_imagematrix,n_sections = 10, height_size=100)

#> $Mean
#> [1] 45.28
#> 
#> $SD
#> [1] 21.53802
#> 
#> $Size
#>  [1] 84.4 56.0 53.6 54.0 54.8 54.4 37.6 24.8 23.2 10.0
#> 

# Conclusions:
# i)  the mean height of the highest black pixel is 45.28 cm.
# ii) standard deviation of highest black height is 21.54.