Cumulative denseness for each line

Proportion of black pixel below each matrix line.

heigh_propotion(imagematrix)

Arguments

imagematrix
The matrix to be analysed.

References

Zehm et al 2003 Multiparameter analysis of vertical vegetation structure based on digital image processing. Flora-Morphology, Distribution, Functional Ecology of Plants, 198: 142-160.

See also

Examples

# First, get a matrix from your image. Here an example of a bush image is used.
bush<-system.file("extdata/bush.JPG",package ="bwimage")
bush_imagematrix<-threshold_color(bush,  "jpeg",  "proportional", compress_rate = 0.1)

# Proportion of black pixel below each matrix line.
heigh_propotion(bush_imagematrix)

#>   [1] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>   [7] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [13] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [19] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [25] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [31] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>  [37] 1.00000000 1.00000000 1.00000000 1.00000000 0.99986117 0.99972234
#>  [43] 0.99944468 0.99944468 0.99944468 0.99944468 0.99944468 0.99944468
#>  [49] 0.99944468 0.99944468 0.99944468 0.99944468 0.99944468 0.99944468
#>  [55] 0.99944468 0.99944468 0.99944468 0.99944468 0.99944468 0.99944468
#>  [61] 0.99944468 0.99944468 0.99944468 0.99944468 0.99944468 0.99916701
#>  [67] 0.99888935 0.99875052 0.99875052 0.99875052 0.99875052 0.99875052
#>  [73] 0.99875052 0.99875052 0.99875052 0.99875052 0.99875052 0.99875052
#>  [79] 0.99875052 0.99875052 0.99875052 0.99875052 0.99875052 0.99875052
#>  [85] 0.99875052 0.99875052 0.99875052 0.99875052 0.99875052 0.99875052
#>  [91] 0.99875052 0.99875052 0.99875052 0.99875052 0.99875052 0.99875052
#>  [97] 0.99875052 0.99875052 0.99875052 0.99875052 0.99875052 0.99861169
#> [103] 0.99847286 0.99833403 0.99833403 0.99833403 0.99833403 0.99819520
#> [109] 0.99819520 0.99819520 0.99819520 0.99805637 0.99791753 0.99777870
#> [115] 0.99763987 0.99666805 0.99500208 0.99291962 0.98986533 0.98625573
#> [121] 0.98222963 0.97709288 0.97237262 0.96765237 0.96362627 0.95835069
#> [127] 0.95349160 0.94891018 0.94418992 0.93849785 0.93322227 0.92725253
#> [133] 0.92183812 0.91697904 0.91253644 0.90934333 0.90615022 0.90254061
#> [139] 0.89893100 0.89448841 0.88949049 0.88463140 0.88088297 0.87768985
#> [145] 0.87449674 0.87047064 0.86519506 0.85991948 0.85478273 0.84992364
#> [151] 0.84575871 0.84187144 0.83895599 0.83659586 0.83381924 0.83145911
#> [157] 0.82896016 0.82604470 0.82257393 0.81910315 0.81577121 0.81257809
#> [163] 0.80924615 0.80508122 0.80105512 0.79772317 0.79286408 0.78717201
#> [169] 0.77995280 0.77190060 0.76551437 0.75912814 0.75329724 0.74635569
#> [175] 0.73969180 0.73344440 0.72678051 0.72081077 0.71511870 0.70817715
#> [181] 0.70165209 0.69540469 0.68818548 0.68068860 0.67263640 0.66444537
#> [187] 0.65667083 0.64889629 0.64098292 0.63237540 0.62404554 0.61391087
#> [193] 0.60322088 0.59128141 0.57948077 0.56754130 0.55560183 0.54380119
#> [199] 0.53269471 0.52214355 0.51173122 0.50229071 0.49215605 0.48063307
#> [205] 0.46883243 0.45508816 0.44148272 0.42704429 0.41552131 0.40621963
#> [211] 0.39886159 0.39192003 0.38553381 0.37859225 0.36998473 0.35901708
#> [217] 0.34804942 0.33860891 0.32902957 0.32153269 0.31431348 0.30764959
#> [223] 0.30001388 0.29251701 0.28488130 0.27752325 0.26988755 0.26086353
#> [229] 0.25267250 0.24545328 0.23851173 0.23143135 0.22421213 0.21685409
#> [235] 0.20866306 0.20033319 0.19158684 0.18325698 0.17284465 0.16298764
#> [241] 0.15285298 0.14230182 0.13188949 0.11967236 0.10620575 0.09093433
#> [247] 0.07552409 0.05858670 0.04109399 0.02137998