Holes description in columns sections

Summary information of holes in a given number of columns (n_sections). n_sections must be set as a number, in this situation hole_columm will sample columns, and apply hole_section_data function for each section. Next, all results will be display on hole_columm output. Example of how column sample works: in a matrix of 250x250 if n_sections = 5 , it will sample columns 1,51,101,151, and 201 and analyze it. In other words, the sections will be following sections of the original matrix [1:250,1] , [1:250,51], [1:250,101], [1:250,151], [1:250,201]. The default for parameter n_sections is "all", it will calculate hole_section_data for each column of pixel. In other words, it will break the image in a number of section equal to the image pixel width.

hole_columm(imagematrix, color = 0, n_sections = "all")

Arguments

imagematrix
The matrix to be analysed.

color
Color of the hole (0 or 1).

n_sections
Sample this number of columns.

Value

N

Number of sections.

Mean

Mean sections size.

SD

standard deviations of sections size.

Min

Minimum sections size sections size.

Max

Maximum sections size.

LH

Stratum with largest hole count.

References

Zehm et al 2003 Multiparameter analysis of vertical vegetation structure based on digital image processing. Flora-Morphology, Distribution, Functional Ecology of Plants, 198: 142-160.

See also

Examples

# First, get a matrix from your image. Here an example of a bush image is used.
bush<-system.file("extdata/bush.JPG",package ="bwimage")
bush_imagematrix<-threshold_color(bush,  "jpeg",  "proportional", compress_rate = 0.1)

# Information of white (i.e. 0s in matrix) holes in 5 columns uniformly sample among matrix.
hole_columm(bush_imagematrix, n_sections=5 )

#> $N
#> [1]  2  2 11  6  3
#> 
#> $Mean
#> [1] 125.00000 124.00000  17.00000  32.50000  79.66667
#> 
#> $SD
#> [1] 161.22035 142.83557  34.27827  65.08994 102.44185
#> 
#> $Min
#> [1] 11 23  2  2  8
#> 
#> $Max
#> [1] 239 225 119 165 197
#> 
#> $LH
#> [1] 3
#> 

# Information of black (i.e. 1s in matrix) holes in 20 columns uniformly sample among matrix.
hole_columm(bush_imagematrix, n_sections=20 )

#> $N
#>  [1]  2  2  2  1  2  5  7  5 15 12 12  8  5  6  5  3  3  4  1  2
#> 
#> $Mean
#>  [1] 125.00000 125.00000 125.00000 249.00000 123.00000  47.60000  31.14286
#>  [8]  43.80000  15.13333  15.16667  13.25000  24.62500  35.20000  33.00000
#> [15]  44.20000  79.66667  74.33333  59.75000 250.00000 124.50000
#> 
#> $SD
#>  [1] 161.22035  86.26703  26.87006        NA 155.56349  50.91463  41.57094
#>  [8]  56.49956  30.20612  31.09175  32.42929  62.38575  73.67971  62.94760
#> [15]  70.26521 109.91057 102.05064  89.33971        NA 147.78532
#> 
#> $Min
#>  [1]  11  64 106 249  13   8   2  10   2   2   2   2   2   2   2   6  10   4 250
#> [20]  20
#> 
#> $Max
#>  [1] 239 186 144 249 233 131 118 144 120 113 116 179 167 161 168 206 192 192 250
#> [20] 229
#> 
#> $LH
#> [1] 9
#>